feat(infra): support renzo warp route for checker tooling (#4353)
### Description

- enable checker tooling to check renzo warp route

### Drive-by changes
- pass `mailboxAddress` to `moduleMatchesConfig`, needed for
`FALLBACK_ROUTER` checks

<!--
Are there any minor or drive-by changes also included?
-->

### Testing

Manual
