+++ typescript/infra/config/environments/mainnet3/warp/renzo-ezETH-addresses-v3.json
@@ -1,0 +1,32 @@
+{
+  "arbitrum": {
+    "xERC20": "0xB26bBfC6d1F469C821Ea25099017862e7368F4E8"
+  },
+  "optimism": {
+    "xERC20": "0xacEB607CdF59EB8022Cc0699eEF3eCF246d149e2"
+  },
+  "base": {
+    "xERC20": "0x2552516453368e42705D791F674b312b8b87CD9e"
+  },
+  "blast": {
+    "xERC20": "0x486b39378f99f073A3043C6Aabe8666876A8F3C5"
+  },
+  "bsc": {
+    "xERC20": "0xE00C6185a5c19219F1FFeD213b4406a254968c26"
+  },
+  "mode": {
+    "xERC20": "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087"
+  },
+  "linea": {
+    "xERC20": "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087"
+  },
+  "ethereum": {
+    "xERC20Lockbox": "0xC59336D8edDa9722B4f1Ec104007191Ec16f7087"
+  },
+  "fraxtal": {
+    "xERC20": "0x3aE8635A4D581d40a6Edfb3f2ED480f9532994F5"
+  },
+  "zircuit": {
+    "xERC20": "0x2552516453368e42705D791F674b312b8b87CD9e"
+  }
+}
