+++ typescript/infra/scripts/check/check-utils.ts
@@ -268,7 +268,7 @@ const getGovernor = async (
         multiProvider,
       );
 
-    // log error and return is foreign deployment chain is specifically checked
+    // log error and return if foreign deployment chain is specifically checked
     if (
       (chain && foreignDeployments[chain]) ||
       (fork && foreignDeployments[fork])
