+++ typescript/sdk/src/ism/utils.ts
@@ -267,7 +267,13 @@ export async function moduleMatchesConfig(
       // check if the mailbox matches the config for fallback routing
       if (config.type === IsmType.FALLBACK_ROUTING) {
         const client = MailboxClient__factory.connect(moduleAddress, provider);
-        const mailboxAddress = await client.mailbox();
+        let mailboxAddress;
+        try {
+          mailboxAddress = await client.mailbox();
+        } catch (error) {
+          matches = false;
+          break;
+        }
         matches =
           matches &&
           mailbox !== undefined &&
@@ -307,7 +313,14 @@ export async function moduleMatchesConfig(
       for (const subModule of subModules) {
         const subModuleMatchesConfig = await Promise.all(
           config.modules.map((c) =>
-            moduleMatchesConfig(chain, subModule, c, multiProvider, contracts),
+            moduleMatchesConfig(
+              chain,
+              subModule,
+              c,
+              multiProvider,
+              contracts,
+              mailbox,
+            ),
           ),
         );
         // The submodule returned by the ISM must match exactly one
@@ -439,7 +452,9 @@ export async function routingModuleDelta(
         contracts,
         mailbox,
       );
-      if (!subModuleMatches) delta.domainsToEnroll.push(originDomain);
+      if (!subModuleMatches) {
+        delta.domainsToEnroll.push(originDomain);
+      }
     }
   }
   return delta;
